+++ lib/src/blocs/boolean_field/boolean_field_state.dart
@@ -11,6 +11,7 @@ class BooleanFieldBlocState extends FieldBlocState<bool?> {
     required super.rules,
     required super.error,
     required super.enabled,
+    required super.loading,
     super.extraData,
     super.formBloc,
   });
@@ -25,6 +26,7 @@ class BooleanFieldBlocState extends FieldBlocState<bool?> {
     Set<ValidationType>? rules,
     Object? error = empty,
     bool? enabled,
+    bool? loading,
     Object? extraData = empty,
     Object? formBloc = empty,
   }) {
@@ -39,6 +41,7 @@ class BooleanFieldBlocState extends FieldBlocState<bool?> {
       rules: rules ?? this.rules,
       error: error == empty ? this.error : error as String?,
       enabled: enabled ?? this.enabled,
+      loading: loading ?? this.loading,
       extraData: extraData == empty ? this.extraData : extraData,
       formBloc: formBloc == empty ? this.formBloc : formBloc as FormBloc?,
     );