+++ lib/src/blocs/date_time_field/date_time_field_state.dart
@@ -11,6 +11,7 @@ class DateTimeFieldBlocState extends FieldBlocState<DateTime?> {
     required super.rules,
     required super.error,
     required super.enabled,
+    required super.loading,
     super.extraData,
     super.formBloc,
     required this.dateFormat,
@@ -32,6 +33,7 @@ class DateTimeFieldBlocState extends FieldBlocState<DateTime?> {
     Set<ValidationType>? rules,
     Object? error = empty,
     bool? enabled,
+    bool? loading,
     Object? extraData = empty,
     Object? formBloc = empty,
     Object? dateFormat = empty,
@@ -49,6 +51,7 @@ class DateTimeFieldBlocState extends FieldBlocState<DateTime?> {
       rules: rules ?? this.rules,
       error: error == empty ? this.error : error as String?,
       enabled: enabled ?? this.enabled,
+      loading: loading ?? this.loading,
       extraData: extraData == empty ? this.extraData : extraData,
       formBloc: formBloc == empty ? this.formBloc : formBloc as FormBloc?,
       dateFormat: