+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+## 3.0.3
+- Added loading field to the FieldBlocState class
+- Added setters - value, loading, enabled, and required to the SingleFieldBloc class
+
 ## 3.0.2
 - Added customSubmit field to the FormBloc class